+++ moor_generator/lib/src/backends/plugin/services/outline.dart
@@ -16,12 +16,12 @@ class MoorOutlineContributor implements OutlineContributor {
     if (moorRequest.isMoorAndParsed) {
       final visitor = _OutlineVisitor(moorRequest, collector);
 
-      moorRequest.parsedMoor.parsedFile.accept(visitor);
+      moorRequest.parsedMoor.parsedFile.acceptWithoutArg(visitor);
     }
   }
 }
 
-class _OutlineVisitor extends RecursiveVisitor<void> {
+class _OutlineVisitor extends RecursiveVisitor<void, void> {
   final MoorRequest request;
   final OutlineCollector collector;
 
@@ -40,33 +40,33 @@ class _OutlineVisitor extends RecursiveVisitor<void> {
   }
 
   @override
-  void visitCreateTableStatement(CreateTableStatement e) {
+  void visitCreateTableStatement(CreateTableStatement e, void arg) {
     _startElement(ElementKind.CLASS, e.tableName, e);
-    super.visitChildren(e);
+    super.visitChildren(e, arg);
     collector.endElement();
   }
 
   @override
-  void visitColumnDefinition(ColumnDefinition e) {
+  void visitColumnDefinition(ColumnDefinition e, void arg) {
     // we use parameters instead of returnType because VS Code doesn't show
     // the return type but we'd really like it to be shown
     _startElement(ElementKind.FIELD, e.columnName, e).parameters = e.typeName;
 
-    super.visitChildren(e);
+    super.visitChildren(e, arg);
     collector.endElement();
   }
 
   @override
-  void visitMoorFile(MoorFile e) {
+  void visitMoorFile(MoorFile e, void arg) {
     _startElement(ElementKind.LIBRARY, request.file.shortName, e);
-    super.visitChildren(e);
+    super.visitChildren(e, arg);
     collector.endElement();
   }
 
   @override
-  void visitMoorDeclaredStatement(DeclaredStatement e) {
+  void visitMoorDeclaredStatement(DeclaredStatement e, void arg) {
     if (!e.isRegularQuery) {
-      super.visitChildren(e);
+      super.visitChildren(e, arg);
       return;
     }
 
@@ -85,7 +85,7 @@ class _OutlineVisitor extends RecursiveVisitor<void> {
       element.parameters = parameterBuilder.toString();
     }
 
-    super.visitChildren(e);
+    super.visitChildren(e, arg);
     collector.endElement();
   }
 }