+++ sqlparser/lib/src/analysis/steps/prepare_ast.dart
@@ -8,19 +8,19 @@ part of '../analysis.dart';
 ///   statement.
 /// - reports syntactic errors that aren't handled in the parser to keep that
 ///   implementation simpler.
-class AstPreparingVisitor extends RecursiveVisitor<void> {
+class AstPreparingVisitor extends RecursiveVisitor<void, void> {
   final List<Variable> _foundVariables = [];
   final AnalysisContext context;
 
   AstPreparingVisitor({this.context});
 
   void start(AstNode root) {
-    root.accept(this);
+    root.accept(this, null);
     _resolveIndexOfVariables();
   }
 
   @override
-  void visitSelectStatement(SelectStatement e) {
+  void visitSelectStatement(SelectStatement e, void arg) {
     // a select statement can appear as a sub query which has its own scope, so
     // we need to fork the scope here. There is one special case though:
     // Select statements that appear as a query source can't depend on data
@@ -81,11 +81,11 @@ class AstPreparingVisitor extends RecursiveVisitor<void> {
       }
     }
 
-    visitChildren(e);
+    visitChildren(e, arg);
   }
 
   @override
-  void visitResultColumn(ResultColumn e) {
+  void visitResultColumn(ResultColumn e, void arg) {
     if (e is StarResultColumn) {
       // doesn't need special treatment, star expressions can't be referenced
     } else if (e is ExpressionResultColumn) {
@@ -93,11 +93,11 @@ class AstPreparingVisitor extends RecursiveVisitor<void> {
         e.scope.register(e.as, e);
       }
     }
-    visitChildren(e);
+    visitChildren(e, arg);
   }
 
   @override
-  void visitQueryable(Queryable e) {
+  void visitQueryable(Queryable e, void arg) {
     final scope = e.scope;
     e.when(
       isTable: (table) {
@@ -120,25 +120,25 @@ class AstPreparingVisitor extends RecursiveVisitor<void> {
       },
     );
 
-    visitChildren(e);
+    visitChildren(e, arg);
   }
 
   @override
-  void visitCommonTableExpression(CommonTableExpression e) {
+  void visitCommonTableExpression(CommonTableExpression e, void arg) {
     e.scope.register(e.cteTableName, e);
-    visitChildren(e);
+    visitChildren(e, arg);
   }
 
   @override
-  void visitNumberedVariable(NumberedVariable e) {
+  void visitNumberedVariable(NumberedVariable e, void arg) {
     _foundVariables.add(e);
-    visitChildren(e);
+    visitChildren(e, arg);
   }
 
   @override
-  void visitNamedVariable(ColonNamedVariable e) {
+  void visitNamedVariable(ColonNamedVariable e, void arg) {
     _foundVariables.add(e);
-    visitChildren(e);
+    visitChildren(e, arg);
   }
 
   void _resolveIndexOfVariables() {
@@ -179,12 +179,12 @@ class AstPreparingVisitor extends RecursiveVisitor<void> {
   }
 
   @override
-  void visitChildren(AstNode e) {
+  void visitChildren(AstNode e, void arg) {
     // hack to fork scopes on statements (selects are handled above)
     if (e is Statement && e is! SelectStatement) {
       _forkScope(e);
     }
 
-    super.visitChildren(e);
+    super.visitChildren(e, arg);
   }
 }