+++ sqlparser/CHANGELOG.md
@@ -1,3 +1,7 @@
+## unreleased
+
+- Added a argument type and argument to the visitor classes
+
 ## 0.5.0
 - Optionally support the `json1` module
 - Optionally support the `fts5` module